refactor: move type definition to a specific file
