+++ src/typedefs.ts
@@ -1,0 +1,37 @@
+export const typeDefs = `#graphql
+  enum Color {
+    ORANGE
+    BLACK
+    RED
+    BLUE
+  }
+
+  type Picture {
+    source: String
+    product: Product
+  }
+
+  type Category {
+    name: String
+    products: [Product]
+  }
+
+  type Product {
+    id: String
+    name: String
+    description: String
+    price: String
+    categories: [Category]
+    picture: Picture
+    color: Color
+  }
+
+  type Query {
+    products: [Product]
+    productsByDescription(productDescription: String): [Product]
+    productsByCategory(categoryName: String): [Product]
+    productById(productId: String): Product,
+    categories: [Category]
+    pictures: [Picture]
+  }
+`;
